moved solvers to solvers.py
